Move code to fix build
