+++ src/System.Diagnostics.DiagnosticSource/src/System/Diagnostics/DiagnosticSource.cs
@@ -60,57 +60,5 @@ namespace System.Diagnostics
         {
             return IsEnabled(name);
         }
-
-        /// <summary>
-        /// Optional: If an instumentation site creating an new activity that was caused
-        /// by something outside the process (e.g. an incomming HTTP request), then that site
-        /// will want to make a new activity and transfer state from that incoming request
-        /// to the activity.   To the extent possible this should be done by the instrumenation
-        /// site (because it is a contract between Activity and the incomming request logic
-        /// at the instrumenation site.   However the instrumenation site can't handle policy
-        /// (for example if sampling is done exactly which requests should be sampled) For this
-        /// the instrumentation site needs to call back out to the logging system and ask it to
-        /// resolve policy (e.g. decide if the Activity's 'sampling' bit should be set)  This
-        /// is what OnActivityImport is for.   It is given the activity as well as a payload
-        /// object that represents the incomming request.   The DiagnosticSource's subscribers
-        /// then have the opportunity to update this activity as desired.   
-        /// 
-        /// Note that this callout is rarely used at instrumentation sites (only those sites
-        /// that are on the 'boundry' of the process), and the instrumetation site will implement
-        /// some default policy (it sets the activity in SOME way), and so this method does not
-        /// need to be overriden if that default policy is fine.   Thus this is call should 
-        /// be used rare (but often important) cases.   
-        /// 
-        /// Note that the type of 'payloadObject' is typed as object here, but for any 
-        /// particular instrumentation site and the subscriber will know the type of 
-        /// the payload and thus cast it and decode it if it needs to. 
-        /// </summary>
-        public virtual void OnActivityImport(Activity activity, object payloadObj) { }
-
-        /// <summary>
-        /// Optional: If an instumentation site is at a location where activities leave the
-        /// process (e.g. an outgoing HTTP request), then that site will want to transfer state 
-        /// from the activity to the outgoing request.    To the extent possible this should be
-        /// done by the instrumenationsite (because it is a contract between Activity and the 
-        /// ougoing request logic at the instrumenation site.   However the instrumenation site 
-        /// can't handle policy (for example whether activity information should be disabled, 
-        /// or written in a older format for compatibility reasons).   For this
-        /// the instrumentation site needs to call back out to the logging system and ask it to
-        /// resolve policy.  This is what OnActivityExport is for.   It is given the activity as
-        /// well as a payloay object that represents the outgoing request.   The DiagnosticSource's 
-        /// subscriber then have the ability to update the outgoing request before it is sent.   
-        /// 
-        /// Note that this callout is rarely used at instrumentation sites (only those sites
-        /// that are on an outgoing 'boundry' of the process).   Moreover typically the default
-        /// policy that the instrumenation site performs (transfer all activity state in a 
-        /// particular outgoing convention), is likely to be fine.   This is only for cases
-        /// where that is a problem.  Thus this is call should be used very rarely and is
-        /// mostly here for symetry with OnActivityImport and future-proofing.  
-        /// 
-        /// Note that the type of 'payloadObject' is typed as object here, but for any 
-        /// particular instrumentation site and the subscriber should know the type of 
-        /// the payload and thus cast it and decode it if it needs to.
-        /// </summary>
-        public virtual void OnActivityExport(Activity activity, object payloadObj) { }
     }
 }
